raspi_bcvdollar_check/requirements.txt
beautifulsoup4==4.13.3
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
colorzero==2.0
distro==1.8.0
font-hanken-grotesk==0.0.2
font-intuitive==0.0.4
font-roboto==0.0.1
font-source-sans-pro==0.0.1
gpiod==2.2.4
gpiodevice==0.0.5
gpiozero==2.0.1
idna==3.3
inky==2.0.0
Jinja2==3.1.5
lgpio==0.2.2.0
lxml==5.3.2
MarkupSafe==3.0.2
numpy==2.2.3
pdoc==15.0.1
pigpio==1.78
pillow==11.1.0
pycryptodomex==3.11.0
Pygments==2.19.1
requests==2.28.1
RPi.GPIO==0.7.1a4
six==1.16.0
smbus2==0.4.2
sn3218==1.2.7
soupsieve==2.6
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
typing_extensions==4.13.1
urllib3==1.26.12
